Improve abstraction
