+++ pkg/wrapper/wrapper.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"fmt"
 	"os"
+
+	"github.com/anttiharju/vmatch/pkg/locate"
 )
 
 type wrapperInterface interface {
@@ -11,7 +13,7 @@ type wrapperInterface interface {
 	Exit(code int)
 	ExitWithPrint(code int, msg string)
 	ExitWithPrintln(code int, msg string)
-	GenerateInstallPath(version string) error
+	GenerateInstallPath(filename string, parse parser, validate validator) error
 }
 
 type Interface interface {
@@ -40,7 +42,35 @@ func (w *BaseWrapper) ExitWithPrintln(exitCode int, message string) {
 	os.Exit(exitCode)
 }
 
-func (w *BaseWrapper) GenerateInstallPath(version string) error {
+type parser func(content []byte) (string, error)
+
+type validator func(version string) (string, error)
+
+func version(filename string, parse parser, validate validator) (string, error) {
+	location, err := locate.File(filename)
+	if err != nil {
+		return "", fmt.Errorf("cannot find version file '%s': %w", filename, err)
+	}
+
+	content, err := os.ReadFile(location)
+	if err != nil {
+		return "", fmt.Errorf("cannot read version file '%s': %w", location, err)
+	}
+
+	version, err := parse(content)
+	if err != nil {
+		return "", fmt.Errorf("could not parse %s: %w", location, err)
+	}
+
+	return validate(version)
+}
+
+func (w *BaseWrapper) GenerateInstallPath(filename string, parse parser, validate validator) error {
+	version, err := version(filename, parse, validate)
+	if err != nil {
+		return fmt.Errorf("failed to get version: %w", err)
+	}
+
 	homeDir, err := os.UserHomeDir()
 	if err != nil {
 		return fmt.Errorf("failed to get install path: %w", err)
